fix: move the auto-playable-film rule into its own module
Review found the appended block had orphaned pickStatusLine's doc comment,
and repeated a false claim that release.ts re-states the same OMDb values
(it type-imports OmdbType directly, erased at build time). Moving the rule
to its own zero-import src/util module, matching resultSort.ts and
resultFilter.ts, fixes both and matches releasePick.ts ranking an
already-chosen release vs. this deciding whether a row is playable at all.
