+++ src/util/releasePick.test.ts
@@ -280,21 +280,6 @@ describe("rankReleases", () => {
 
 import { pickStatusLine } from "./releasePick";
 
-import { autoPlayableFilm } from "./releasePick";
-
-describe("autoPlayableFilm", () => {
-  it("trusts OMDb's medium over the filter", () => {
-    expect(autoPlayableFilm("movie", "all")).toBe(true);
-    expect(autoPlayableFilm("series", "movie")).toBe(false);
-  });
-
-  it("falls back to the filter when OMDb said nothing", () => {
-    expect(autoPlayableFilm(null, "movie")).toBe(true);
-    expect(autoPlayableFilm(undefined, "all")).toBe(false);
-    expect(autoPlayableFilm(null, "tv")).toBe(false);
-  });
-});
-
 describe("pickStatusLine", () => {
   it("names the release when the preference was met", () => {
     const pick = pickBestRelease([c("Kestrel.2010.1080p.BluRay.x264")], NO_PREFS, FILM)!;
